refactor(models): split scope.rs, and drop what nothing calls
scope.rs held seven types and a stray `impl Definition`. Four modules now hold one
idea each:

  scope             the tree, and which scope a position sits in
  symbol_table      declarations indexed by name, beside that tree
  registry          what a traversal collected, before anything is resolved
  analysis_context  everything one traversal produced

`Definition::new_simple` goes to definition.rs, where the type lives.

Fifteen public methods went rather than moved — 136 lines that nothing calls, in
three cases a whole chain of them: `SymbolTable::lookup_symbol` was the only caller
of `ScopeTree::lookup_symbol_in_scope_chain`, and `CodeAnalysisContext::lookup_symbol_in_scope`
the only caller of `DefinitionRegistry::lookup_definition`. Neither root had a
caller of its own. Both registries also carried a full type-parameter API that
nothing ever read.

Behaviour is unchanged: 743 of 743, precision and recall 1.000, identical to the
recorded baseline.
