+++ crates/core/src/models/mod.rs
@@ -1,3 +1,4 @@
+pub mod analysis_context;
 pub mod ast_traverser;
 pub mod definition;
 pub mod dependency;
@@ -6,11 +7,14 @@ pub mod language;
 pub mod metrics;
 pub mod module;
 pub mod position;
+pub mod registry;
 pub mod scope;
+pub mod symbol_table;
 pub mod type_system;
 pub mod usage;
 
 // Re-export all public types for convenient access
+pub use analysis_context::CodeAnalysisContext;
 pub use ast_traverser::{ASTScopeTraverser, NodeDefinitionExtractor, NodeUsageExtractor};
 pub use definition::{Accessibility, Definition, DefinitionType, ScopeId};
 pub use dependency::{Dependency, DependencyType};
@@ -19,9 +23,8 @@ pub use language::Language;
 pub use metrics::{AnalysisResult, LineMetrics, OverallAnalysisReport};
 pub use module::{ImportInfo, ImportType, Module, ModuleId, ModuleTree, Visibility};
 pub use position::Position;
-pub use scope::{
-    CodeAnalysisContext, DefinitionRegistry, Scope, ScopeTree, ScopeType, SymbolEntry, SymbolTable,
-    UsageRegistry,
-};
+pub use registry::{DefinitionRegistry, UsageRegistry};
+pub use scope::{Scope, ScopeTree, ScopeType};
+pub use symbol_table::{SymbolEntry, SymbolTable};
 pub use type_system::{InferenceContext, Type};
 pub use usage::{Usage, UsageKind};
